pyproject.toml
[project]
name = "tribal-mapper"
version = "0.1.0"
description = "AI Tribal Knowledge Mapper — indexes codebases into structured context files and routes context to developers at query time."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.34.0",
]

[tool.ruff]
# Target version for syntax and bug-finds
target-version = "py311"
line-length = 88

[tool.ruff.lint]
# "E" (pycodestyle), "F" (Pyflakes), "I" (isort) are the essentials
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["D203"]                        # Example: ignore specific docstring rules

[tool.ruff.lint.isort]
# Ensure imports are grouped nicely
combine-as-imports = true
force-single-line = false

[tool.ruff.format]
# Use double quotes (standard Black style)
quote-style = "double"
indent-style = "space"
